CHANGE: Finished implementing Administrator Endpoints. CHANGE: Added ConfirmedException when trying to update or delete a confirmed Entry. ongoing development: Category Controller and linking to Entry.

+++ src/main/java/com/danilojakob/m223/punchclock/service/EntryService.java
@@ -1,6 +1,7 @@
 package com.danilojakob.m223.punchclock.service;
 
 import com.danilojakob.m223.punchclock.domain.ApplicationUser;
+import com.danilojakob.m223.punchclock.exceptions.ConfirmedException;
 import com.danilojakob.m223.punchclock.repository.EntryRepository;
 import com.danilojakob.m223.punchclock.domain.Entry;
 import org.springframework.stereotype.Service;
@@ -27,11 +28,33 @@ public class EntryService {
         return entryRepository.findByUsername(username);
     }
 
-    public void deleteById(Long id) throws Exception {
+    public void deleteById(Long id) throws ConfirmedException {
+        // Check if entry is already confirmed
+        if (checkIfEntryIsConfirmed(id)) {
+            throw new ConfirmedException("Entry already Confirmed");
+        }
         entryRepository.deleteById(id);
     }
 
-    public Entry updateEntry(Entry entry) throws Exception {
+    /**
+     * Confirm an entry
+     * @param id {@link Long} id of the entry
+     * @return {@link Entry} updated entry
+     */
+    public Entry confirmEntry(Long id) {
+        Entry entry = entryRepository.getOne(id);
+        entry.setConfirmed(true);
         return entryRepository.saveAndFlush(entry);
+    }
+
+    public Entry updateEntry(Entry entry) throws ConfirmedException {
+        if (checkIfEntryIsConfirmed(entry.getId())) {
+            throw new ConfirmedException("Entry is already confirmed");
+        }
+        return entryRepository.saveAndFlush(entry);
+    }
+
+    public boolean checkIfEntryIsConfirmed(Long id) {
+        return entryRepository.getOne(id).getConfirmed();
     }
 }
